Providerでテーマ変数を保存する機能追加に合わせて、Storyのテーマ設定を修正
